+++ turboui/src/WorkMap/components/WorkMapTable.tsx
@@ -1,11 +1,32 @@
-import WorkMap from "..";
-import classNames from "../../../utils/classnames";
+import { WorkMap } from "..";
+import { TableRow } from "./TableRow";
+import classNames from "../../utils/classnames";
 
 interface Props {
+  items: WorkMap.Item[];
   filter: WorkMap.Filter;
 }
 
-export function TableHeader({ filter }: Props) {
+export function WorkMapTable({ items, filter }: Props) {
+  return (
+    <div className="overflow-x-auto bg-surface-base rounded-b-lg">
+      <table className="min-w-full divide-y divide-surface-outline">
+        <TableHeader filter={filter} />
+        <tbody>
+          {items.map((item, idx) => (
+            <TableRow key={item.id} item={item} level={0} isLast={idx === items.length - 1} filter={filter} />
+          ))}
+        </tbody>
+      </table>
+    </div>
+  );
+}
+
+interface HeaderProps {
+  filter: WorkMap.Filter;
+}
+
+export function TableHeader({ filter }: HeaderProps) {
   const isCompletedPage = filter === "completed";
 
   return (
